Refactor to handle multiple application property files
The main function has been significantly refactored to handle multiple application  property files instead of just one. The aim is to improve flexibility and scalability of property file management. Now, the program reads all yaml files in a directory, flattens their data and compares it with environment-specific master config. This commits also includes the creation of functions to gather all yaml files from a directory and to get application properties from the files.